move shortcuts to their own file
